[Feature]: 초기설정 SplashView, MapView생성 완료
